feat: big changes

+++ server/auth/index.ts
@@ -1,4 +1,7 @@
+'use server'
+
 import type { Session, User } from '@prisma/client'
+import { cookies } from 'next/headers'
 import { sha256 } from '@oslojs/crypto/sha2'
 import { encodeBase32LowerCaseNoPadding, encodeHexLowerCase } from '@oslojs/encoding'
 
@@ -6,7 +9,8 @@ import { db } from '@/server/db'
 
 const EXPIRES_IN = 1000 * 60 * 60 * 24 * 30
 
-export function generateSessionToken(): string {
+// eslint-disable-next-line @typescript-eslint/require-await
+export const generateSessionToken = async (): Promise<string> => {
   const bytes = new Uint8Array(20)
   crypto.getRandomValues(bytes)
   const token = encodeBase32LowerCaseNoPadding(bytes)
@@ -23,7 +27,10 @@ export const createSession = async (token: string, userId: string): Promise<Sess
   return await db.session.create({ data: session })
 }
 
-export const validateSessionToken = async (token: string): Promise<SessionValidation> => {
+export const auth = async (): Promise<SessionValidation> => {
+  const token = (await cookies()).get('auth_token')?.value ?? ''
+  if (!token) return { expires: new Date(Date.now()) }
+
   const sessionToken = encodeHexLowerCase(sha256(new TextEncoder().encode(token)))
   const result = await db.session.findUnique({
     where: { sessionToken },
@@ -48,7 +55,10 @@ export const validateSessionToken = async (token: string): Promise<SessionValida
   return { user, expires: session.createdAt }
 }
 
-export const invalidateSession = async (token: string): Promise<void> => {
+export const invalidateSession = async (): Promise<void> => {
+  const token = (await cookies()).get('auth_token')?.value ?? ''
+  if (!token) return
+
   const sessionToken = encodeHexLowerCase(sha256(new TextEncoder().encode(token)))
   await db.session.delete({ where: { sessionToken } })
 }
